+++ .eslintrc.js
@@ -18,11 +18,17 @@ module.exports = {
 
 	extends: 'standard',
 
+	plugins: ['prettier'],
+
 	rules: {
 		indent: ['error', 'tab', { ignoredNodes: ['TemplateLiteral > *'] }],
 		'no-tabs': 0,
 		'no-console': 0,
-		semi: [1, 'always']
+		semi: [1, 'always'],
+		'space-before-function-paren': [
+			'error',
+			{ anonymous: 'never', named: 'never', asyncArrow: 'always' }
+		]
 	},
 
 	globals: {